+++ CHANGELOG.md
@@ -34,6 +34,7 @@ The format is based on [Keep a Changelog]
   - Update ECP admin claim verifier qualifications eligibility to use dates
     based on qualification type
   - Add new DQT fields to admin claim verifier qualifications notes
+  - Add qualification name to admin ECP claim qualification task presenter
 - All policies
   - Home Address selection
     - Add home address search screen to claim journey